move ConfigFilesProvider to core module
