+++ cfg4j-core/src/main/java/org/cfg4j/source/ConfigFilesProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cfg4j.source.git;
+package org.cfg4j.source;
 
 import java.io.File;
 import java.nio.file.Path;
