+++ cfg4j-core/src/test/java/org/cfg4j/source/DefaultConfigFilesProviderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cfg4j.source.git;
+package org.cfg4j.source;
 
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
